companion for 411, Update stack runner block_coinbase
