+++ bin/node/runtime/pangolin/src/lib.rs
@@ -663,7 +663,7 @@ impl_runtime_apis! {
 		}
 
 		fn author() -> H160 {
-			<dvm_ethereum::Pallet<Runtime>>::find_author()
+			<darwinia_evm::Pallet<Runtime>>::find_author()
 		}
 
 		fn storage_at(address: H160, index: U256) -> H256 {
