+++ frame/evm/src/tests.rs
@@ -17,9 +17,9 @@
 // along with Darwinia. If not, see <https://www.gnu.org/licenses/>.
 
 // --- std ---
-use std::{collections::BTreeMap, str::FromStr};
+use std::{collections::BTreeMap, marker::PhantomData, str::FromStr};
 // --- substrate ---
-use frame_support::{assert_ok, traits::GenesisBuild};
+use frame_support::{assert_ok, traits::GenesisBuild, ConsensusEngineId};
 use frame_system::mocking::*;
 use sp_core::H256;
 use sp_runtime::{
@@ -29,11 +29,13 @@ use sp_runtime::{
 };
 // --- darwinia ---
 use crate::{self as darwinia_evm, runner::stack::Runner, *};
+use sp_consensus_aura::sr25519::AuthorityId as AuraId;
 
 type Block = MockBlock<Test>;
 type UncheckedExtrinsic = MockUncheckedExtrinsic<Test>;
 
 type Balance = u64;
+type Aura = pallet_aura::Module<Test>;
 
 darwinia_support::impl_test_account_data! {}
 
@@ -99,12 +101,26 @@ impl pallet_timestamp::Config for Test {
 	type WeightInfo = ();
 }
 
+impl pallet_aura::Config for Test {
+	type AuthorityId = AuraId;
+}
+
 /// Fixed gas price of `0`.
 pub struct FixedGasPrice;
 impl FeeCalculator for FixedGasPrice {
 	fn min_gas_price() -> U256 {
 		// Gas price is always one token per gas.
 		0.into()
+	}
+}
+
+pub struct FindAuthorTruncated<F>(PhantomData<F>);
+impl<F: FindAuthor<u32>> FindAuthor<H160> for FindAuthorTruncated<F> {
+	fn find_author<'a, I>(_digests: I) -> Option<H160>
+	where
+		I: 'a + IntoIterator<Item = (ConsensusEngineId, &'a [u8])>,
+	{
+		Some(H160::from_str("1234500000000000000000000000000000000000").unwrap())
 	}
 }
 
@@ -169,6 +185,7 @@ impl Config for Test {
 
 	type AddressMapping = ConcatAddressMapping<Self::AccountId>;
 	type BlockHashMapping = crate::SubstrateBlockHashMapping<Self>;
+	type FindAuthor = FindAuthorTruncated<Aura>;
 
 	type RingCurrency = Ring;
 	type KtonCurrency = Kton;
@@ -258,5 +275,16 @@ fn fail_call_return_ok() {
 			U256::default(),
 			None,
 		));
+	});
+}
+
+#[test]
+fn find_author() {
+	new_test_ext().execute_with(|| {
+		let author = EVM::find_author();
+		assert_eq!(
+			author,
+			H160::from_str("1234500000000000000000000000000000000000").unwrap()
+		);
 	});
 }
